fixed the obs issue by flattening it and sending it as a single list of  values

+++ Envs/Complicated_Env1.py
@@ -28,7 +28,7 @@ class GameEnv(Env):
         self.render_mode = render_mode
 
         # self.observation_space = Box(low=0, high=self.screen_width, shape=(360,))
-        self.observation_space = Box(low=np.zeros(360), high=self.screen_width * np.ones(360), dtype=np.float32)
+        self.observation_space = Box(low=np.zeros(360, dtype=np.float32), high=self.screen_width * np.ones(360, dtype=np.float32), dtype=np.float32)
         # defining the observation and action spaces for all the agents
 
         # defining the action space based on total number of predator and prey
